pyproject.toml
[project]
name = "django-saas-auth"
version = "0.4.0"
description = "The authentication module for building a SaaS product with Django"
authors = [
    { name = "Hsiaoming Yang", email = "me@lepture.com" }
]
dependencies = [
    "Django",
    "otpauth",
    "djangorestframework",
    "django-saas-base>=0.14.0",
]
readme = "README.md"
requires-python = ">= 3.12"
license = {text = "BSD-3-Clause"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Framework :: Django :: 6.0",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "django-stubs",
    "drf_spectacular",
    "pytest",
    "pytest-django",
    "pytest-cov",
    "python-dotenv",
]
lint = [
    "ruff",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/saas_auth"]

[tool.hatch.build.targets.sdist]
exclude = [
    ".python-version",
    ".github/*",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"

[tool.pytest.ini_options]
addopts = "--no-migrations"
pythonpath = ["src"]
testpaths = ["tests"]
filterwarnings = ["error"]
DJANGO_SETTINGS_MODULE = "tests.settings"

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["saas_auth"]
omit = [
    "*/migrations/*",
    "*/admin.py",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "@(abc\\.)?abstractmethod",
]
